Renamed `evaluation` to `metrics`.
